feat(monorepo): add and configure prettier
Prettier is an opinionated code formatter. Any formatting done by Prettier should never affect the meaning of the code, so Prettier should be able to run automatically in the background. And since its opinionated, only one convention is needed for the monorepo. For these reasons, Prettier is configured globally at the monorepo root.