refactor(packages/vitest-config): re-export defineconfig and mergeconfig (#32)
Vitest is hoisted to the global npm space to alleviate
the need to have to import vitest and other related
dependencies in each and every workspace package. But
the config in the workspaces needs to import
defineConfig and mergeConfig from vitest and there is
an ESLint rule to prevent imports of deps not present
in the workspaces package.json. So this refactor simply
re-exports defineConfig and mergeConfig out of
@toolchain/vitest-config.