fix(dockerfile): turbo build --filter
The `...` on the --filter was on the wrong side. Having
on the left side builds the app and whatever is
dependent on the app, which in most cases is nothing.
What we want is to build the app and all projects that
the app depends on. So move the `...` to the right
side.