fix(packages/api): return the token via the signup call
Setting the Cookie in the API doesn't
work when the API is running outside the
web application.

Should not set cookies in the API layer as
it is a higher-order concern. So just
return the token from the API call and let
the upstream caller decide what to do with
it.