fix(toolchain/eslint-config): svelte attr sorting
Was using perfectionist's eslint rule to
sort svelte attributes but this was causing
conflicts with prettier svelte plugin. Could
not disable sorting in prettier so decided
to revert back to the sorting rules in
svelte's eslint plugin which doesn't seem to
conflict with svelte's prettier sorting.

This was an example of the issue inside a svelte component:
<svelte:element  class="foo" this={undefined} />