feat(toolchain/eslint-config): revamp modular configuration approach for node profile
Original configuration was created at a time when
most of the eslint plugins were not supporting flat
config. As a result, there was some clever ways to get
around mixing legacy config with flat config. This new
approach utilized the newer flat configs provided by
the plugins themselves while still keeping with the
modularity of the configs, housing the specific plugin
config to individual files.