chore(monorepo): update monorepo and workspace dependencies to latest (#417)
update to eslint-plugin-svelte triggered new violation
`svelte/no-navigation-without-resolve` resolve using
best practices. Note: in packages/ui-lib-svelte the rule
itself was ignored in the button component for now as 
not seeing any good way to make this work in such a 
generic fashion.

needed to turn off @typescript-eslint/dot-notation as 
there is this catch 22 that if there in an active .env file
you get one behavior and if there iisn't and active .env
file you get a different behavior.